유광현 - 주문 관리 기능구현 중 1245

+++ naedam2/src/main/java/com/naedam/mir9/order/model/vo/Order.java
@@ -19,7 +19,7 @@ public class Order implements Serializable {
 	 */
 	private static final long serialVersionUID = 1L;
 	
-	private int orderNo;
+	private long orderNo;
 	private int memberNo;
 	private int productNo;
 	private int orderStatusNo;
